Update s3-table-engine with a non-AWS example
I recent was helped in the issue queue with mounting a wasabi bucket as a disk, but had difficulty troubleshooting the error I was getting when mapping the AWS style URLs to the Wasabi style (https://github.com/ClickHouse/ClickHouse/issues/42043#issuecomment-1309638255). I added a small example for wasabi (and likely other users) that explains that both the bucket and the folder need to be added to the end of the endpoint path.