Update mutations.md
Fix typo in project documentation:

This commit addresses a minor error in the project documentation. In the ALTER TABLE statement example, the DELETE keyword was omitted. The correct statement should be ALTER TABLE clicks ON CLUSTER main_cluster DELETE WHERE visit_date < '2022-01-02 15:00:00' AND page_id = '573'. I've updated the documentation to reflect this change.