fix: Replace some it's with its
Sometimes "it's" is correct. To check, replace "it's" with "it is". If
the sentance still makes sense, "it's" is totally fine! If it doesn't,
"its" is correct.

This PR replaces some incorrect "it's" with "its".

I noticed this while reading this specific docs page. I scanned the repo
for all uses of "it's" expecting to find more errors, but these were the
only ones I found!