Update cloud-quick-start.md
I have removed all screenshots of gettingStarted. In a couple of weeks, we will have statistically significant data to see if we want to deploy to 100% of users. For now, this guide walks users on only using SQL console. In future we will push more ClickPipes content here too. Will update this fully after A/B test.