fix: Duplicate materialized view definition in `Use Cases and Caveats`
There are 2 example under `Use Cases and Caveats` with the first showing a non-optimum materialized view join and the second showing a join which filters on the changed rows. The 2 code blocks are identical, just slightly differently formatted in their white space.

This changes the initial query to what should be the non-optimum version.